backend/requirements.txt
openai
requests
python-decouple